Ładowanie i wyświetlanie modeli

+++ Vulkan3DEngine/Src/GraphicsEngine/Renderer/Buffers/IndexBuffer/IndexBuffer.h
@@ -6,7 +6,7 @@
 class IndexBuffer : public Buffer
 {
 public:
-	IndexBuffer(std::vector<uint16_t> indices, Renderer* renderer);
+	IndexBuffer(std::vector<uint32_t> indices, Renderer* renderer);
 	~IndexBuffer();
 	
 	void bind() override;
